refactor: design builder as web component
