+++ global.d.ts
@@ -1,0 +1,9 @@
+export {};
+
+declare global {
+	interface Window {
+		styleguideCustomizeData?: unknown;
+		styleguideDesignTokenLibrary?: unknown;
+		styleguideCustomizeInitMode?: unknown;
+	}
+}
