+++ source/design-builder/app/adapters/createFullPageModeAdapter.ts
@@ -1,0 +1,8 @@
+import type { DesignBuilderModeAdapter } from '../../root/types';
+import { initializeFullPageDesignBuilder } from '../../modes/full-page/initializeFullPageDesignBuilder';
+
+export function createFullPageModeAdapter(): DesignBuilderModeAdapter {
+	return ({ hostElement, configuration }) => {
+		initializeFullPageDesignBuilder(hostElement, configuration.tokenData);
+	};
+}
